Added relationship User-Idea

+++ src/idea/idea.service.ts
@@ -1,35 +1,69 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { HttpErrorFilter } from 'src/shared/http-error.filter';
+import { UserEntity } from 'src/user/user.entity';
 import { Repository } from 'typeorm';
-import { ideaDto } from './dto/ideaDto';
+import { ideaDto, ideaRO } from './dto/ideaDto';
 import { IdeaEntity } from './idea.entity';
 
 @Injectable()
 export class IdeaService {
 
-    constructor(@InjectRepository(IdeaEntity) private ideaRepository: Repository<IdeaEntity>) { }
+    constructor(@InjectRepository(IdeaEntity) private ideaRepository: Repository<IdeaEntity>,
+        @InjectRepository(UserEntity) private userRepository: Repository<UserEntity>) { }
 
-    async findAll() {
-        return await this.ideaRepository.find();
+
+    private toResponseObject(idea: IdeaEntity): ideaRO {
+        return { ...idea, author: idea.author.toResponseObject(false) };
     }
 
-    async createOne(idea: ideaDto) {
-        const createIdea = await this.ideaRepository.create(idea);
-        await this.ideaRepository.save(createIdea)
-        return createIdea;
+    async findAll(userId): Promise<ideaRO[]> {
+        // const ideas = await this.ideaRepository.find({ relations: ['author'] });
+        const ideas = await this.ideaRepository.find({  where: { author: userId }, relations: ['author'] });
+        return ideas.map(idea => this.toResponseObject(idea));
     }
 
-    async readOne(id) {
-        return await this.ideaRepository.findOne({ where: { id } });
+    async createOne(userId: string, data: ideaDto): Promise<ideaRO> {
+        const user = await this.userRepository.findOne({ where: { id: userId } });
+        const idea = await this.ideaRepository.create({ ...data, author: user });
+        await this.ideaRepository.save(idea)
+        return this.toResponseObject(idea);
     }
 
-    async updateOne(id, idea) {
-        await this.ideaRepository.update({ id }, idea);
-        return await this.ideaRepository.findOne({ id });
+    async readOne(id): Promise<ideaRO> {
+        const idea = await this.ideaRepository.findOne({ where: { id }, relations: ['author'] });
+        if (!idea) {
+            throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+        }
+        return this.toResponseObject(idea);
     }
 
-    async deleteOne(id) {
+    async updateOne(id, userId, data): Promise<ideaRO> {
+        let idea = await this.ideaRepository.findOne({ where: { id }, relations: ['author'] })
+
+        if (!idea) {
+            throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+        }
+        this.ensureOwnership(idea, userId);
+        await this.ideaRepository.update({ id }, data);
+        idea = await this.ideaRepository.findOne({ where: { id }, relations: ['author'] });
+        return this.toResponseObject(idea);
+    }
+
+    async deleteOne(id: string, userId: string) {
+        let idea = await this.ideaRepository.findOne({ where: { id }, relations: ['author'] })
+        if (!idea) {
+            throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+        }
+        this.ensureOwnership(idea, userId);
         await this.ideaRepository.delete({ id });
-        return {deleted : true};
+        return this.toResponseObject(idea);
+    }
+
+    // Helper Function to ensure Creator & Updator/Deletor is same
+    private ensureOwnership(idea: IdeaEntity, userId: string) {
+        if (idea.author.id !== userId) {
+            throw new HttpException('Incorrect user', HttpStatus.UNAUTHORIZED);
+        }
     }
 }
